Modified simpat, tabpat, and sitepat. The new versions use "long double" to accumulate branch lengths associated with different site patterns. I was worried about losing precision when adding small values to a large sum.
Added code to mapmix, which calculates the global array of site pattern
frequencies and prints these out at the end. This is for debugging. I want to
make sure that the site patterns that mapmix is seeing are the same as those
tabulated by sitepat.

+++ src/simpat.c
@@ -259,7 +259,7 @@ int main(int argc, char **argv) {
            (doSing ? "Including" : "Excluding"));
     printf("# Number of site patterns: %lu\n", npat);
     tipId_t     pat[npat];
-    double      patCount[npat];
+    long double patCount[npat];
     int         lblsize = 100;
     char        lblbuff[lblsize];
     memset(patCount, 0, sizeof(patCount));
@@ -367,7 +367,7 @@ int main(int argc, char **argv) {
     printf("# %13s %20s", "SitePat", "E[count]");
     putchar('\n');
     for(i = 0; i < npat; ++i) {
-        printf("%15s %20.7lf\n",
+        printf("%15s %20.7Lf\n",
                patLbl(lblsize, lblbuff, pat[i], &lndx), patCount[i]);
     }
